requirements.txt
langchain==0.3.17
langchain-openai==0.3.0
langgraph==0.2.68
python-dotenv==1.0.1
langchain-core>=0.3.29,<0.4.0
langgraph-checkpoint==2.1.2
python-docx==1.1.2
streamlit==1.41.1
aiofiles==24.1.0
pydantic==2.10.6
tiktoken==0.8.0
nest_asyncio==1.6.0
